server/requirements.txt
flask
groq
python-dotenv
flask-cors